requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-classy-tags==4.0.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-seed==0.3.1
django-taggit==3.1.0
django-taggit-serializer==0.1.7
django-taggit-templatetags2==1.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==15.3.4
idna==3.4
joblib==1.2.0
lxml==4.9.1
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
toposort==1.7
tzdata==2022.7
urllib3==1.26.13
